+++ CHANGELOG.md
@@ -9,8 +9,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.11.5 / 5.66.5] - 2023-10-
 
+### Changed
+- improved suspend process ahndling [#3375](https://github.com/sandboxie-plus/Sandboxie/issues/3375)
+
 ### Fixed
 - fixed issue with auto updater not offering version updates
+
+
 
 
 ## [1.11.4 / 5.66.4] - 2023-10-06